pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "morphobase"
version = "1.3.0"
description = "A Computational Substrate for Organism-First Adaptive Intelligence"
readme = "README.md"
authors = [{name = "Justin"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "torch>=2.0",
    "PyYAML>=6.0",
    "matplotlib>=3.7",
    "tqdm>=4.65",
    "pandas>=2.0"
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.3"]

[tool.setuptools.packages.find]
include = ["morphobase*"]
